+++ docs/changelog.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## v1.0-beta2 (2021-xx-xx)
+## v1.0-beta2 (2021-04-04)
 
-- New helper function get_runtime_milliseconds
+First improvements after user feedback - thank you Dietmar(@daust):
+
+- Splitting level verbose into debug and trace (we have now error, warning, info, debug and trace)
+- Remove method permanent (is now a attribute in all log methods)
+- Overload all log messages (procedure and function returning log_id)
+- Improved docs to reflect the changes
 
 ## v1.0-beta1 (2021-03-14)